comment create, delete, edit are now all dependent on sessions

+++ backend/src/routes/commentRoute.ts
@@ -1,21 +1,23 @@
-import express, { Request, Response, Router} from "express";
+import express, { Request, Response, Router } from "express";
 
 import User from "../models/user";
 import Comment, { IComment } from "../models/comment";
 import mongoose from "mongoose";
 import Post from "../models/post";
+import { IsLoggedIn } from "../middlewares/authorizedOnly";
 
 const router: Router = express.Router();
 
-router.post("/submitcomment", async (req: Request, res: Response): Promise<any> => {
+router.post("/submitcomment", IsLoggedIn, async (req: Request, res: Response): Promise<any> => {
   try {
-    const { username, body, publicId, parentId, type } = req.body;
+    const { body, publicId, parentId, type } = req.body;
 
     if (!body || !publicId || !type || (type === "Reply" && !parentId)) {
-      return res.status(400).json({ message: "Body, post, type, and parent comment are required."})
+      return res.status(400).json({ message: "Body, post, type, and parent comment are required." })
     }
 
-    const user = await User.findOne({ username: username });
+
+    const user = await User.findOne({ _id: req.session.userId });
 
     if (!user) {
       return res.status(404).json({ message: "User not found." });
@@ -25,11 +27,11 @@ router.post("/submitcomment", async (req: Request, res: Response): Promise<any> 
       const parentComment = await Comment.findOne({ _id: parentId });
 
       if (!parentComment) {
-        return res.status(404).json({ message: "Parent comment not found."});
+        return res.status(404).json({ message: "Parent comment not found." });
       }
 
       const newComment = new Comment({
-        username: username,
+        username: user.username,
         userId: user._id,
         body: body,
         publicId: publicId,
@@ -42,7 +44,8 @@ router.post("/submitcomment", async (req: Request, res: Response): Promise<any> 
       parentComment.replies.push(savedComment._id as mongoose.Types.ObjectId);
       await parentComment.save();
 
-      return res.status(201).json({ message: "Reply created successfully.",
+      return res.status(201).json({
+        message: "Reply created successfully.",
         newReply: {
           commentID: savedComment._id,
           username: savedComment.username,
@@ -51,16 +54,17 @@ router.post("/submitcomment", async (req: Request, res: Response): Promise<any> 
           publicId: savedComment.publicId,
           parentId: savedComment.parentId,
           replies: savedComment.replies,
-        }});
+        }
+      });
     } else if (type === "Comment") {
       const parentPost = await Post.findOne({ publicId: parentId });
 
       if (!parentPost) {
-        return res.status(404).json({ message: "Parent post not found."});
+        return res.status(404).json({ message: "Parent post not found." });
       }
 
       const newComment = new Comment({
-        username: username,
+        username: user.username,
         userId: user._id,
         body: body,
         publicId: publicId,
@@ -72,7 +76,8 @@ router.post("/submitcomment", async (req: Request, res: Response): Promise<any> 
       parentPost.comments.push(savedComment._id as mongoose.Types.ObjectId);
       await parentPost.save();
 
-      return res.status(201).json({ message: "Comment created successfully",
+      return res.status(201).json({
+        message: "Comment created successfully",
         newComment: {
           commentID: savedComment._id,
           username: savedComment.username,
@@ -89,13 +94,18 @@ router.post("/submitcomment", async (req: Request, res: Response): Promise<any> 
   }
 });
 
-router.post("/editcomment/:id", async (req: Request, res: Response): Promise<any> => {
+router.post("/editcomment/:id", IsLoggedIn, async (req: Request, res: Response): Promise<any> => {
   const { body } = req.body;
 
   const comment = await Comment.findOne({ _id: req.params.id });
 
   if (!comment) {
     return res.status(404).json({ message: "Comment not found." });
+  }
+
+  if (comment.userId.toString() !== req.session.userId) {
+    res.status(403).json({ message: "This is not your comment" });
+    return;
   }
 
   comment.body = body;
@@ -114,13 +124,19 @@ router.post("/editcomment/:id", async (req: Request, res: Response): Promise<any
   });
 })
 
-router.post("/deletecomment/:id", async (req: Request, res: Response): Promise<any> => {
+router.post("/deletecomment/:id", IsLoggedIn, async (req: Request, res: Response): Promise<any> => {
   try {
-    const comment = await Comment.findOne({_id: req.params.id});
+    const comment = await Comment.findOne({ _id: req.params.id });
 
     if (!comment) {
       return res.status(404).json({ message: "Comment not found." });
     }
+
+    if (comment.userId.toString() !== req.session.userId) {
+      res.status(403).json({ message: "Not your comment!" });
+      return;
+    }
+
 
     await Promise.all(comment.replies.map(async (replyId) => {
       await Comment.deleteOne({ _id: replyId });
@@ -134,7 +150,7 @@ router.post("/deletecomment/:id", async (req: Request, res: Response): Promise<a
     } else {
       await Post.updateOne(
         { publicId: comment.publicId },
-        { $pull: { comments: comment._id }}
+        { $pull: { comments: comment._id } }
       );
     }
 
@@ -153,7 +169,8 @@ router.get("/retrievecomment/:id", async (req: Request, res: Response): Promise<
       return res.status(404).json({ message: "Comment not found." });
     }
 
-    return res.status(200).json({ message: "Comment successfully pulled.",
+    return res.status(200).json({
+      message: "Comment successfully pulled.",
       comment: {
         commentID: data._id,
         username: data.username,
@@ -162,7 +179,8 @@ router.get("/retrievecomment/:id", async (req: Request, res: Response): Promise<
         publicId: data.publicId,
         parentId: data.parentId,
         replies: data.replies,
-      }});
+      }
+    });
   } catch (err: unknown) {
     return res.status(500).json({ message: err });
   }
